+++ tests/testthat/test-read-lines.R
@@ -13,6 +13,12 @@ test_that("read_lines handles embedded nuls", {
   expect_equal(read_lines("null-file", progress = FALSE), c("a,b,c", "1,2,", "3,4,5"))
 })
 
+test_that("read_lines uses na argument", {
+  expect_equal(read_lines("sample_text.txt", na = "abc", progress = FALSE), c(NA_character_, "123"))
+  expect_equal(read_lines("sample_text.txt", na = "123", progress = FALSE), c("abc", NA_character_))
+  expect_equal(read_lines("sample_text.txt", na = c("abc", "123"), progress = FALSE), c(NA_character_, NA_character_))
+})
+
 # These tests are slow so are commented out
 #test_that("long vectors are supported", {
   #tmp <- tempfile(fileext = ".gz")